implement modulo operation for memory addresses
Before, when loading and storing to or from an adress, a natural number
was used. This simplified the semantics, but allowed reading and writing
to memory locations just after the allowed address range if a large word
needed to be read from or stored in multiple memory locations. The
original idea was that the caller needs to take care of this via eg.
alignment checks.

After the chat with Roberto yesterday, I implemented the modulo
operation on memory addresses directly in bir.