introduce extra status BST_AssertionViolated
Before, the status was BST_failed, if an assertion got violated. While
BST_AssertionViolated is still a failure for e.g. Hoare-triples, it
allows to reason about programs more fine-grained. We can now
distinguish that a program fails, because an assertion is violated
(which is kind of "expected" sometimes (otherwise you could remove the
assertions) or whether the program fails because it is not well-formed
in some sense (which should be avoided). In particular this allows us to
show that the programs produced by the program lifter only fail, if an
assertion is violated.