start clarification of when a program fails
This developments tries to make it more explicit when a bir program
returns the status code "BST_Failed". However, before this real goal can
be even started, some infrastructure about the variables used by a
program is needed. This commit introduces the variables used (read and
written) and changed (written) by statements, blocks and programs. It is
then shown that a program at most changes the vars in its read-set and
the execution depends only on the vars in the whole set.

This should provide the background to formally talk about the failed
status.