finer grained preprocessing of step theorems
Make "computation" of memory containing instruction machine specific.
This simplifies the code, there is now now fancy computation needed.
This allows changing the theorem more, by e.g. applying arithmetic
simplifications which would interfer with the fancy computation.

As a result, the code is more rebust, but incidentially also much faster
and simpler. This is a necessary preparation for instantiating the PC
late and for lifting stores.