finer grained multistep execution
For the lifter, the existing definitions for multistep execution turned
out to be incovenient. It is handy to be able to stop when reaching
certain PCs. Moreover, counting them is useful as well.

This commit contains a generalisation of `bir_exec_steps_opt`. In
addition to the old functionality, a predicate that determines which
program counters should be counted is provided. The number of steps is
now with respect to this counting and the resulting number, as well as
the number of steps is returned.

The internal implementation changed considerably. Most obviously, there
is now a Datatype for the result. Before, option types were used to
encode if an execution terminated. Nowe there is a type
`bir_execution_result`.

`(ll, NONE)` becomes `BER_Looping ll` and
`(fromList ol, SOME (c, st))` becomes `BER_Ended ol c c_pc_count st`

Notice that in the case the execution terminates, there is an extra
value `c_pc_count` returned now. It is shown that if an execution loops,
at least one block of the program is executed infinitely often. This
allows to show that if a given maximum step number is given and at least
entering a block is counted, the execution terminates. As a result,
`bir_exec_step_n` and `bir_exec_block_n` can be shown to always
terminate. Therefore, they return a tuple instead of a value of type
`bir_execution_result_t`.

While the internas changed significantly, the syntax and semantics of
`bir_exec_step_n` stayed the same. `bir_exec_steps` only changed
slightly. More statements of lemmata about these functions are
unchanged.