lifter: sort out distinct labels properly
Before, we checked that the address labels were distinct by demanding
that their numeric values are strictly sorted and just used brute force
on the string labels. Since for our current examples, there are no
string labels, this worked fine.

Now, we demand formally that the string labels are prefixed with a
string version of their corresponding address label. This was an
informal requirement before anyhow. By making it formal, we can show
that we don't need to consider the string labels at all. Moreover, the
sorting trick for the address labels has been cleaned up.