lifter: add hex-codes to lifted BIR program
use BL_Address_HC instead of BL_Address in order to add the original
hex-codes of the machine instructions to the lited bir-program.
Semantically, this is a noop, since

BL_Address_HC addr hex_code := BL_Address addr

However, it might be useful for debugging, since the user can interpret

"BL_Adresss addr hc" as "this block is the translation of the machine
instruction with hex-code hc found at address addr in memory".