bir_ppLib: fix issue with Unicode characters
It looks like the HOLstring pretty-printer was invoked to pretty-print
unicode characters (such as /\ or ==>), which obviously failed when
calling `stringSyntax.fromHOLstring`.

I decided to fix this by correctly handling the exception raised by
`stringSyntax.fromHOLstring`. A somewhat better solution would be to
specialize this pretty-printer to only pretty-print BIR strings, such as
labels (that was the initial motivation of this PP). Furthermore, this
would be better wrt the lib's name (BIR PP), because in the current
state, the PP turns ALL strings in green.